backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-filter==23.1
djoser==2.1.0
drf-extra-fields==3.7.0
Pillow==9.0.0
reportlab==4.2.5